pyproject.toml
[tool.poetry]
name = "medixo-back-end"
version = "0.1.0"
description = ""
authors = ["mohammedalsamki <momosamak21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
django-phonenumber-field = "^6.0.0"
django-phonenumbers = "^1.0.1"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
djangorestframework-simplejwt = "^5.0.0"
review = "^1510"
django-cors-headers = "^3.10.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
